+++ GuildWars2.ArenaNet/Model/MapDetails.cs
@@ -23,8 +23,8 @@ namespace GuildWars2.ArenaNet.Model
 
         public string ContinentName { get; set; }
 
-        public List<List<int>> MapRect { get; set; }
+        public List<List<double>> MapRect { get; set; }
 
-        public List<List<int>> ContinentRect { get; set; }
+        public List<List<double>> ContinentRect { get; set; }
     }
 }
